+++ src/index.ts
@@ -12,3 +12,7 @@ export { default as FormikRating } from "./FormikRating/FormikRating";
 export { default as FormikToggleButtonGroup } from "./FormikToggleButtonGroup/FormikToggleButtonGroup";
 export { default as FormikNumberField } from "./FormikNumberField/FormikNumberField";
 export { default as FormikPasswordField } from "./FormikPasswordField/FormikPasswordField";
+export { default as FormikFileUpload } from "./FormikFileUpload/FormikFileUpload";
+export { default as FormikChipInput } from "./FormikChipInput/FormikChipInput";
+export { default as FormikRadioCards } from "./FormikRadioCards/FormikRadioCards";
+export { default as FormikSearchField } from "./FormikSearchField/FormikSearchField";